break up Home component i.e. single responsibility
